Update README.md
How to change application icons in the Pocketbook home screen?
I found a workaround for this problem: My PocketBook (InkPad 3) allows me to change and rearrange the apps on the Home screen, but the settings are not saved. It reverts to the default apps and arrangement after reboot or after a few minutes.